@@ -1,9 +1,9 @@
-name: wordle
+name: Kankerdle
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0315+20220315
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,7 +29,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -45,14 +44,14 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.0
+  icons_launcher: ^2.1.7
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -63,14 +62,14 @@ flutter:
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
   assets:
-    - 'assets/All.txt'
-    - 'assets/CET4.txt'
-    - 'assets/CET6.txt'
-    - 'assets/CET4 + 6.txt'
-    - 'assets/GRE Slim.txt'
-    - 'assets/TOEFL Slim.txt'
-    - 'assets/TOEFL.txt'
-    - 'assets/HighSchool.txt'
+    - "assets/All.txt"
+    - "assets/CET4.txt"
+    - "assets/CET6.txt"
+    - "assets/CET4 + 6.txt"
+    - "assets/GRE Slim.txt"
+    - "assets/TOEFL Slim.txt"
+    - "assets/TOEFL.txt"
+    - "assets/HighSchool.txt"
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
@@ -97,3 +96,16 @@ flutter:
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
+
+icons_launcher:
+  image_path: "assets/icons/logo.png"
+  platforms:
+    android:
+      enable: false
+    ios:
+      enable: false
+    web:
+      enable: true
+      image_path: "assets/icons/logo.png"
+      favicon_path: "assets/icons/logo.png"
+
