@@ -8,14 +8,14 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:wordle/event_bus.dart';
-import 'package:wordle/validation_provider.dart';
-import 'package:wordle/display_pannel.dart';
-import 'package:wordle/instruction_pannel.dart';
-import 'package:wordle/popper_generator.dart';
+import 'package:Kankerdle/event_bus.dart';
+import 'package:Kankerdle/validation_provider.dart';
+import 'package:Kankerdle/display_pannel.dart';
+import 'package:Kankerdle/instruction_pannel.dart';
+import 'package:Kankerdle/popper_generator.dart';
 
 class GamePage extends StatefulWidget {
-  const GamePage({Key? key, required this.database, required this.wordLen, required this.maxChances, required this.gameMode}) : super(key: key);
+  const GamePage({super.key, required this.database, required this.wordLen, required this.maxChances, required this.gameMode});
 
   final Map<String, List<String>> database;
   final int wordLen;
@@ -26,19 +26,17 @@ class GamePage extends StatefulWidget {
   State<GamePage> createState() => _GamePageState();
 }
 
-class _GamePageState extends State<GamePage> with TickerProviderStateMixin{
+class _GamePageState extends State<GamePage> with TickerProviderStateMixin {
   late AnimationController _controller;
 
   void _onGameEnd(dynamic args) {
     var result = args as bool;
-    if(result == true) {
+    if (result == true) {
       _controller.forward().then((v) {
         _controller.reset();
         mainBus.emit(event: "Result", args: result);
-      }
-      );
-    }
-    else {
+      });
+    } else {
       mainBus.emit(event: "Result", args: result);
     }
   }
@@ -73,7 +71,7 @@ class _GamePageState extends State<GamePage> with TickerProviderStateMixin{
                 fontWeight: FontWeight.bold,
                 fontSize: 20.0,
               ),
-              title: const Text('WORDLE'),
+              title: const Text('Kankerdle'),
               centerTitle: true,
               //iconTheme: const IconThemeData(color: Colors.black),
               actions: [
@@ -89,7 +87,8 @@ class _GamePageState extends State<GamePage> with TickerProviderStateMixin{
                       animation: rotateAnimation,
                       builder: (context, child) {
                         return Transform(
-                          transform: Matrix4.rotationZ(rotateAnimation.status == AnimationStatus.reverse ? 2 * pi - rotateAnimation.value : rotateAnimation.value),
+                          transform: Matrix4.rotationZ(
+                              rotateAnimation.status == AnimationStatus.reverse ? 2 * pi - rotateAnimation.value : rotateAnimation.value),
                           alignment: Alignment.center,
                           child: Opacity(
                             opacity: opacAnimation.value,
@@ -109,28 +108,28 @@ class _GamePageState extends State<GamePage> with TickerProviderStateMixin{
                 IconButton(
                   icon: const Icon(Icons.help_outline_outlined),
                   //color: Colors.black,
-                  onPressed: (){
+                  onPressed: () {
                     showInstructionDialog(context: context);
                   },
                 ),
                 IconButton(
                   icon: const Icon(Icons.refresh_rounded),
                   //color: Colors.black,
-                  onPressed: (){
+                  onPressed: () {
                     mainBus.emit(event: "NewGame", args: []);
                   },
                 ),
               ],
             ),
             body: Container(
+              color: Theme.of(context).brightness == Brightness.dark ? Colors.grey[850] : Colors.white,
               child: ValidationProvider(
                 database: widget.database,
                 wordLen: widget.wordLen,
                 maxChances: widget.maxChances,
                 gameMode: widget.gameMode,
-                child: WordleDisplayWidget(wordLen: widget.wordLen, maxChances: widget.maxChances),
+                child: KankerdleDisplayWidget(wordLen: widget.wordLen, maxChances: widget.maxChances),
               ),
-              color: Theme.of(context).brightness == Brightness.dark ? Colors.grey[850] : Colors.white,
             ),
           ),
         ),
@@ -138,7 +137,7 @@ class _GamePageState extends State<GamePage> with TickerProviderStateMixin{
           child: PartyPopperGenerator(
             direction: PopDirection.fowardX,
             motionCurveX: FunctionCurve(func: (t) {
-              return - t * t / 2 + t;
+              return -t * t / 2 + t;
             }),
             motionCurveY: FunctionCurve(func: (t) {
               return 4 / 3 * t * t - t / 3;
@@ -155,7 +154,7 @@ class _GamePageState extends State<GamePage> with TickerProviderStateMixin{
           child: PartyPopperGenerator(
             direction: PopDirection.backwardX,
             motionCurveX: FunctionCurve(func: (t) {
-              return - t * t / 2 + t;
+              return -t * t / 2 + t;
             }),
             motionCurveY: FunctionCurve(func: (t) {
               return 4 / 3 * t * t - t / 3;
