@@ -6,27 +6,27 @@
  */
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:wordle/event_bus.dart';
-import 'package:wordle/loading_page.dart';
-import 'package:wordle/single_selection.dart';
-import 'package:wordle/selection_group.dart';
-import 'package:wordle/scroll_behav.dart';
+import 'package:Kankerdle/event_bus.dart';
+import 'package:Kankerdle/loading_page.dart';
+import 'package:Kankerdle/single_selection.dart';
+import 'package:Kankerdle/selection_group.dart';
+import 'package:Kankerdle/scroll_behav.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:wordle/instruction_pannel.dart';
+import 'package:Kankerdle/instruction_pannel.dart';
 import 'dart:io';
 import 'dart:math';
 
 Future<void> loadSettings() async {
   Directory documentsDirectory = await getApplicationDocumentsDirectory();
   String documentsPath = documentsDirectory.path + Platform.pathSeparator;
-  File settings = File(documentsPath + "settings.txt");
-  if(!(await settings.exists())) {
+  File settings = File("${documentsPath}settings.txt");
+  if (!(await settings.exists())) {
     var defaultSettings = "5\nCET4\nLight";
     settings.writeAsString(defaultSettings);
   }
   List<String> dicBooks = ["validation.txt", "CET4.txt"];
-  for(String dicName in dicBooks) {
-    if(!(await File(documentsPath + dicName).exists())) {
+  for (String dicName in dicBooks) {
+    if (!(await File(documentsPath + dicName).exists())) {
       //Copy file
       ByteData data = await rootBundle.load("assets/CET4.txt");
       List<int> bytes = data.buffer.asUint8List(data.offsetInBytes, data.lengthInBytes);
@@ -35,19 +35,18 @@ Future<void> loadSettings() async {
   }
 }
 
-void main(){
+void main() {
   runApp(const MyApp());
 }
 
 class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
-
   var brightness = Brightness.light;
 
   void _onThemeChange(dynamic args) {
@@ -64,19 +63,20 @@ class _MyAppState extends State<MyApp> {
   }
 
   @override
-  void dispose(){
+  void dispose() {
     mainBus.offBus(event: "ToggleTheme", callBack: _onThemeChange);
     super.dispose();
   }
-  
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       scrollBehavior: MyScrollBehavior(),
-      title: 'Wordle',
+      title: 'Kankerdle',
       theme: ThemeData(
         primarySwatch: Colors.grey,
         brightness: brightness,
+        useMaterial3: true,
       ),
       routes: {
         "/": (context) => const HomePage(),
@@ -87,7 +87,7 @@ class _MyAppState extends State<MyApp> {
 }
 
 class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
+  const HomePage({super.key});
 
   @override
   State<HomePage> createState() => _HomePageState();
@@ -105,7 +105,7 @@ class _HomePageState extends State<HomePage> {
     return FutureBuilder(
       future: readSettings(),
       builder: (context, snapshot) {
-        if(snapshot.connectionState == ConnectionState.done) {
+        if (snapshot.connectionState == ConnectionState.done) {
           var mode = Theme.of(context).brightness;
           return Scaffold(
             body: Align(
@@ -123,7 +123,11 @@ class _HomePageState extends State<HomePage> {
                             alignment: Alignment.bottomLeft,
                             child: Padding(
                               padding: const EdgeInsets.only(left: 30.0, bottom: 10.0),
-                              child: Text('WORDLE', style: TextStyle(fontSize: 30.0, fontWeight: FontWeight.w300, color: mode == Brightness.light ? Colors.grey[850]! : Colors.white)),
+                              child: Text('Kankerdle',
+                                  style: TextStyle(
+                                      fontSize: 30.0,
+                                      fontWeight: FontWeight.bold,
+                                      color: mode == Brightness.light ? Colors.grey[850]! : Colors.white)),
                             ),
                           ),
                           const Spacer(),
@@ -134,8 +138,8 @@ class _HomePageState extends State<HomePage> {
                               child: Row(
                                 children: [
                                   AnimatedSwitcher(
-                                    duration: const Duration(milliseconds: 750),
-                                    reverseDuration: const Duration(milliseconds: 750),
+                                    duration: const Duration(milliseconds: 300),
+                                    reverseDuration: const Duration(milliseconds: 300),
                                     switchInCurve: Curves.bounceOut,
                                     switchOutCurve: Curves.bounceIn,
                                     transitionBuilder: (child, animation) {
@@ -145,7 +149,9 @@ class _HomePageState extends State<HomePage> {
                                         animation: rotateAnimation,
                                         builder: (context, child) {
                                           return Transform(
-                                            transform: Matrix4.rotationZ(rotateAnimation.status == AnimationStatus.reverse ? 2 * pi - rotateAnimation.value : rotateAnimation.value),
+                                            transform: Matrix4.rotationZ(rotateAnimation.status == AnimationStatus.reverse
+                                                ? 2 * pi - rotateAnimation.value
+                                                : rotateAnimation.value),
                                             alignment: Alignment.center,
                                             child: Opacity(
                                               opacity: opacAnimation.value,
@@ -165,12 +171,12 @@ class _HomePageState extends State<HomePage> {
                                   IconButton(
                                     icon: const Icon(Icons.help_outline_outlined),
                                     //color: Colors.black,
-                                    onPressed: (){
+                                    onPressed: () {
                                       showInstructionDialog(context: context);
                                     },
                                   ),
-                                ]
-                              )
+                                ],
+                              ),
                             ),
                           )
                         ],
@@ -178,7 +184,7 @@ class _HomePageState extends State<HomePage> {
                     ),
                     const Expanded(
                       child: Padding(
-                        padding: EdgeInsets.symmetric(horizontal: 20.0),
+                        padding: EdgeInsets.symmetric(horizontal: 8.0),
                         child: OfflinePage(),
                       ),
                     ),
@@ -187,8 +193,7 @@ class _HomePageState extends State<HomePage> {
               ),
             ),
           );
-        }
-        else {
+        } else {
           return const Text('');
         }
       },
@@ -197,29 +202,44 @@ class _HomePageState extends State<HomePage> {
 }
 
 class OfflinePage extends StatefulWidget {
-  const OfflinePage({Key? key}) : super(key: key);
+  const OfflinePage({super.key});
 
   @override
   State<OfflinePage> createState() => _OfflinePageState();
 }
 
 class _OfflinePageState extends State<OfflinePage> {
-  int wordLen = 5;
-  int maxChances = 6;
+  int wordLen = 6;
+  int maxChances = 1;
   int dicBookIndex = 0;
-  var dicBook = [["All", "Full wordlist", "A", Colors.indigo], ["HighSchool", "HighSchool wordlist", "H", Colors.amber],
-                ["CET4", "CET4 wordlist", "4", Colors.green[400]], ["CET6", "CET6 wordlist", "6", Colors.teal[400]],
-                ["CET4 + 6", "CET4&6 wordlist", "46", Colors.teal[600]], ["TOEFL Slim", "TOEFL without CET4&6", "T", Colors.blue[400]],
-                ["TOEFL", "Full TOEFL wordlist", "T", Colors.cyan[400]], ["GRE Slim", "GRE without CET4&6", "G", Colors.pink[200]]];
+  var dicBook = [
+    ["All", "Full wordlist", "A", Colors.indigo],
+    ["HighSchool", "HighSchool wordlist", "H", Colors.amber],
+    ["CET4", "CET4 wordlist", "4", Colors.green[400]],
+    ["CET6", "CET6 wordlist", "6", Colors.teal[400]],
+    ["CET4 + 6", "CET4&6 wordlist", "46", Colors.teal[600]],
+    ["TOEFL Slim", "TOEFL without CET4&6", "T", Colors.blue[400]],
+    ["TOEFL", "Full TOEFL wordlist", "T", Colors.cyan[400]],
+    ["GRE Slim", "GRE without CET4&6", "G", Colors.pink[200]]
+  ];
   late final List<Widget> dicBookSelections;
   var wordLenSelectionColors = [Colors.green[300], Colors.green[500], Colors.teal[300], Colors.teal[500], Colors.pink[300]];
-  var chancesSelectionColors = [Colors.deepOrange[600], Colors.orange[400], Colors.cyan, Colors.blue[400], Colors.blue[600], Colors.teal[400], Colors.teal[600], Colors.green[700]];
+  var chancesSelectionColors = [
+    Colors.deepOrange[600],
+    Colors.orange[400],
+    Colors.cyan,
+    Colors.blue[400],
+    Colors.blue[600],
+    Colors.teal[400],
+    Colors.teal[600],
+    Colors.green[700]
+  ];
 
   @override
   void initState() {
     super.initState();
     dicBookSelections = [
-      for(int i = 0; i < dicBook.length; i++)
+      for (int i = 0; i < dicBook.length; i++)
         generateSelectionBox(
           id: i,
           width: 190,
@@ -236,8 +256,6 @@ class _OfflinePageState extends State<OfflinePage> {
     ];
   }
 
-
-
   @override
   Widget build(BuildContext context) {
     return SizedBox(
@@ -249,54 +267,7 @@ class _OfflinePageState extends State<OfflinePage> {
           children: [
             Container(
               width: double.infinity,
-              height: 150.0,
-              decoration: BoxDecoration(
-                color: wordLenSelectionColors[wordLen - 4]!.withOpacity(0.2),
-                borderRadius: BorderRadius.circular(10.0),
-              ),
-              alignment: Alignment.centerLeft,
-              padding: const EdgeInsets.all(10.0),
-              margin: const EdgeInsets.symmetric(horizontal: 10.0, vertical: 10.0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.fromLTRB(15.0, 10.0, 0.0, 10.0),
-                    child: Text(
-                      'Word Length',
-                      style: TextStyle(
-                        fontSize: 22.0,
-                        fontWeight: FontWeight.bold,
-                        color: wordLenSelectionColors[wordLen - 4]!,
-                      ),
-                    ),
-                  ),
-                  SelectionGroupProvider(
-                    defaultSelection: 5,
-                    onChanged: (sel) => setState(() => wordLen = sel),
-                    selections: SingleChildScrollView(
-                      scrollDirection: Axis.horizontal,
-                      child: Row(
-                        children: [
-                          for(int i = 4; i <= 8; i++)
-                            generateSelectionBox(
-                              id: i,
-                              width: 80.0,
-                              height: 80.0,
-                              color: wordLenSelectionColors[i - 4]!,
-                              primaryText: '$i',
-                              primaryTextSize: 25.0,
-                            ),
-                        ],
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            Container(
-              width: double.infinity,
-              height: 150.0,
+              height: 151.0,
               decoration: BoxDecoration(
                 color: chancesSelectionColors[maxChances - 1]!.withOpacity(0.2),
                 borderRadius: BorderRadius.circular(10.0),
@@ -319,13 +290,13 @@ class _OfflinePageState extends State<OfflinePage> {
                     ),
                   ),
                   SelectionGroupProvider(
-                    defaultSelection: 6,
+                    defaultSelection: 1,
                     onChanged: (sel) => setState(() => maxChances = sel),
                     selections: SingleChildScrollView(
                       scrollDirection: Axis.horizontal,
                       child: Row(
                         children: [
-                          for(int i = 1; i <= 8; i++)
+                          for (int i = 1; i <= 8; i++)
                             generateSelectionBox(
                               id: i,
                               width: 80.0,
@@ -335,41 +306,6 @@ class _OfflinePageState extends State<OfflinePage> {
                               primaryTextSize: 25.0,
                             ),
                         ],
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            Container(
-              width: double.infinity,
-              decoration: BoxDecoration(
-                color: (dicBook[dicBookIndex][3]! as Color).withOpacity(0.2),
-                borderRadius: BorderRadius.circular(10.0),
-              ),
-              margin: const EdgeInsets.all(10.0),
-              padding: const EdgeInsets.all(15.0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.fromLTRB(15.0, 10.0, 0.0, 10.0),
-                    child: Text(
-                      'Word List',
-                      style: TextStyle(
-                        fontSize: 22.0,
-                        fontWeight: FontWeight.bold,
-                        color: (dicBook[dicBookIndex][3]! as Color),
-                      ),
-                    ),
-                  ),
-                  SelectionGroupProvider(
-                    defaultSelection: 0,
-                    onChanged: (sel) => setState(() => dicBookIndex = sel),
-                    selections: SingleChildScrollView(
-                      scrollDirection: Axis.horizontal,
-                      child: Row(
-                        children: dicBookSelections,
                       ),
                     ),
                   ),
@@ -388,7 +324,7 @@ class _OfflinePageState extends State<OfflinePage> {
                       ),
                       margin: const EdgeInsets.all(10.0),
                       child: InkWell(
-                        onTap: (){
+                        onTap: () {
                           Navigator.of(context).push(MaterialPageRoute(builder: (context) {
                             return LoadingPage(dicName: dicBook[dicBookIndex][0] as String, wordLen: wordLen, maxChances: maxChances, gameMode: 0);
                           }));
@@ -398,7 +334,7 @@ class _OfflinePageState extends State<OfflinePage> {
                           alignment: Alignment.center,
                           padding: const EdgeInsets.symmetric(horizontal: 30.0, vertical: 30.0),
                           child: const Text(
-                            'Start Normal',
+                            'Start',
                             style: TextStyle(
                               fontSize: 22.0,
                               fontWeight: FontWeight.bold,
