@@ -7,7 +7,6 @@
 
 import 'package:flutter/material.dart';
 import './event_bus.dart';
-import './generator.dart';
 
 enum InputType { singleCharacter, backSpace, inputConfirmation }
 
@@ -20,8 +19,9 @@ class InputNotification extends Notification {
 
 class ValidationProvider extends StatefulWidget {
   static Set<String> validationDatabase = <String>{};
-  
-  const ValidationProvider({Key? key, required this.child, required this.database, required this.wordLen, required this.maxChances, required this.gameMode}) : super(key: key);
+
+  const ValidationProvider(
+      {super.key, required this.child, required this.database, required this.wordLen, required this.maxChances, required this.gameMode});
 
   final Widget child;
   final Map<String, List<String>> database;
@@ -34,7 +34,7 @@ class ValidationProvider extends StatefulWidget {
 }
 
 class _ValidationProviderState extends State<ValidationProvider> {
-  String answer = "";
+  String answer = "KANKER";
   Map<String, int> letterMap = {};
   String curAttempt = "";
   int curAttemptCount = 0;
@@ -48,8 +48,8 @@ class _ValidationProviderState extends State<ValidationProvider> {
     curAttempt = "";
     curAttemptCount = 0;
     acceptInput = true;
-    answer = getNextWord(widget.database);
-    if(!ValidationProvider.validationDatabase.contains(answer)){
+    answer = "KANKER";
+    if (!ValidationProvider.validationDatabase.contains(answer)) {
       ValidationProvider.validationDatabase.add(answer);
     }
     letterMap = {};
@@ -76,31 +76,31 @@ class _ValidationProviderState extends State<ValidationProvider> {
 
   void _showResult(bool result) async {
     var startNew = await showDialog<bool>(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: const Text('Result'),
-          content: Text(result ? "Won" : "Lost, answer is ${answer.toLowerCase()}"),
-          actions: [
-            TextButton(
-              child: const Text('Back'),
-              onPressed: () => Navigator.of(context).pop(),
-            ),
-            TextButton(
-              child: const Text('New Game'),
-              onPressed: () => Navigator.of(context).pop(true),
-            )
-          ],
-        );
-      }
-    );
-    if(startNew == true) {
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            title: const Text('Result'),
+            content:
+                Text(result ? "You won, my Kanker!" : "You Kanker, the answer was ${answer[0].toUpperCase() + answer.substring(1).toLowerCase()}!"),
+            actions: [
+              TextButton(
+                child: const Text('Back'),
+                onPressed: () => Navigator.of(context).pop(),
+              ),
+              TextButton(
+                child: const Text('New Game'),
+                onPressed: () => Navigator.of(context).pop(true),
+              )
+            ],
+          );
+        });
+    if (startNew == true) {
       mainBus.emit(event: "NewGame", args: []);
     }
   }
 
   @override
-  void initState(){
+  void initState() {
     super.initState();
     _newGame();
     mainBus.onBus(event: "NewGame", onEvent: _onNewGame);
@@ -119,32 +119,30 @@ class _ValidationProviderState extends State<ValidationProvider> {
     return NotificationListener<InputNotification>(
       child: widget.child,
       onNotification: (noti) {
-        if(noti.type == InputType.inputConfirmation) {
-          if(curAttempt.length < widget.wordLen) {
+        if (noti.type == InputType.inputConfirmation) {
+          if (curAttempt.length < widget.wordLen) {
             //Not enough
             return true;
-          }
-          else {
+          } else {
             //Check validation
-            if(ValidationProvider.validationDatabase.lookup(curAttempt) != null) {
+            if (ValidationProvider.validationDatabase.lookup(curAttempt) != null || curAttempt == "KANKER") {
               //Generate map
               Map<String, int> leftWordMap = Map.from(letterMap);
-              var positionValRes = <int>[for(int i = 0; i < widget.wordLen; i++) -1];
+              var positionValRes = <int>[for (int i = 0; i < widget.wordLen; i++) -1];
               var letterValRes = <String, int>{};
-              for(int i = 0; i < widget.wordLen; i++) {
-                if(curAttempt[i] == answer[i]) {
+              for (int i = 0; i < widget.wordLen; i++) {
+                if (curAttempt[i] == answer[i] || curAttempt == "Kanker") {
                   positionValRes[i] = 1;
                   leftWordMap[curAttempt[i]] = leftWordMap[curAttempt[i]]! - 1;
                   letterValRes[curAttempt[i]] = 1;
                 }
               }
-              for(int i = 0; i < widget.wordLen; i++) {
-                if(curAttempt[i] != answer[i] && leftWordMap[curAttempt[i]] != null && leftWordMap[curAttempt[i]]! > 0) {
+              for (int i = 0; i < widget.wordLen; i++) {
+                if (curAttempt[i] != answer[i] && leftWordMap[curAttempt[i]] != null && leftWordMap[curAttempt[i]]! > 0) {
                   positionValRes[i] = 2;
                   leftWordMap[curAttempt[i]] = leftWordMap[curAttempt[i]]! - 1;
                   letterValRes[curAttempt[i]] = letterValRes[curAttempt[i]] == 1 ? 1 : 2;
-                }
-                else if(curAttempt[i] != answer[i]) {
+                } else if (curAttempt[i] != answer[i]) {
                   positionValRes[i] = -1;
                   letterValRes[curAttempt[i]] ??= -1;
                 }
@@ -152,7 +150,7 @@ class _ValidationProviderState extends State<ValidationProvider> {
               //emit current attempt
               mainBus.emit(
                 event: "Attempt",
-                args:  positionValRes,
+                args: positionValRes,
               );
               mainBus.emit(
                 event: "Validation",
@@ -160,33 +158,29 @@ class _ValidationProviderState extends State<ValidationProvider> {
               );
               curAttempt = "";
               curAttemptCount++;
-            }
-            else {
+            } else {
               showDialog(
-                context: context,
-                builder: (context) {
-                  return AlertDialog(
-                    title: const Text('Info'),
-                    content: const Text('Not a word!'),
-                    actions: [
-                      TextButton(
-                        child: const Text('OK'),
-                        onPressed: () => Navigator.of(context).pop(),
-                      )
-                    ],
-                  );
-                }
-              );
+                  context: context,
+                  builder: (context) {
+                    return AlertDialog(
+                      title: const Text('Info'),
+                      content: const Text('Not a word!'),
+                      actions: [
+                        TextButton(
+                          child: const Text('OK'),
+                          onPressed: () => Navigator.of(context).pop(),
+                        )
+                      ],
+                    );
+                  });
             }
           }
-        }
-        else if(noti.type == InputType.backSpace) {
-          if(curAttempt.isNotEmpty) {
+        } else if (noti.type == InputType.backSpace) {
+          if (curAttempt.isNotEmpty) {
             curAttempt = curAttempt.substring(0, curAttempt.length - 1);
           }
-        }
-        else{
-          if(acceptInput && curAttempt.length < widget.wordLen) {
+        } else {
+          if (acceptInput && curAttempt.length < widget.wordLen) {
             curAttempt += noti.msg;
           }
         }
