@@ -11,7 +11,7 @@ import './group_shared.dart';
 
 class SingleSelectionBox extends StatefulWidget {
   const SingleSelectionBox({
-    Key? key,
+    super.key,
     required this.id,
     required this.unselectedChild,
     required this.selectedChild,
@@ -25,8 +25,7 @@ class SingleSelectionBox extends StatefulWidget {
     this.borderRadius = 0.0,
     this.borderColor = Colors.transparent,
     this.borderWidth = 0.0,
-  }):
-  super(key: key);
+  });
 
   final int id;
   final Widget unselectedChild;
@@ -45,7 +44,7 @@ class SingleSelectionBox extends StatefulWidget {
   State<SingleSelectionBox> createState() => _SingleSelectionBoxState();
 }
 
-class _SingleSelectionBoxState extends State<SingleSelectionBox> with TickerProviderStateMixin{
+class _SingleSelectionBoxState extends State<SingleSelectionBox> with TickerProviderStateMixin {
   bool selected = false;
 
   //Record last pressed shift from center point
@@ -62,17 +61,17 @@ class _SingleSelectionBoxState extends State<SingleSelectionBox> with TickerProv
     super.initState();
     _scaleController = AnimationController(
       vsync: this,
-      duration: const Duration(milliseconds: 750),
+      duration: const Duration(milliseconds: 300),
     );
     _opacController = AnimationController(
       vsync: this,
-      duration: const Duration(milliseconds: 750),
+      duration: const Duration(milliseconds: 300),
     );
     _shrinkController = AnimationController(
       vsync: this,
-      duration: const Duration(milliseconds: 500),
+      duration: const Duration(milliseconds: 300),
     );
-    if(context.findAncestorWidgetOfExactType<GroupSharedData>()!.selected == widget.id) {
+    if (context.findAncestorWidgetOfExactType<GroupSharedData>()!.selected == widget.id) {
       _scaleController.value = _scaleController.upperBound;
       selected = true;
     }
@@ -82,7 +81,7 @@ class _SingleSelectionBoxState extends State<SingleSelectionBox> with TickerProv
   void didChangeDependencies() {
     super.didChangeDependencies();
     var newData = context.findAncestorWidgetOfExactType<GroupSharedData>()!;
-    if(newData.selected != widget.id && selected) {
+    if (newData.selected != widget.id && selected) {
       _opacController.forward();
       selected = false;
     }
@@ -92,131 +91,130 @@ class _SingleSelectionBoxState extends State<SingleSelectionBox> with TickerProv
   Widget build(BuildContext context) {
     var data = GroupSharedData.of(context)!;
     //Use gesture detector to detect the position pressed for animations
-    return LayoutBuilder(
-      builder:(context, constraints) {
-        var decoration = BoxDecoration(
-          boxShadow: [BoxShadow(
+    return LayoutBuilder(builder: (context, constraints) {
+      var decoration = BoxDecoration(
+        boxShadow: [
+          BoxShadow(
             color: data.selected == widget.id ? widget.selectedShadowColor : widget.unselectedShadowColor,
             blurRadius: widget.shadowBlurRadius,
             spreadRadius: widget.shadowSpreadRadius,
-          )],
-          color: widget.unselectedBackgroundColor,
-          border: widget.borderWidth > 0 ? Border.all(
-            color: data.selected == widget.id ? widget.selectedBackgroundColor : widget.borderColor,
-            width: widget.borderWidth
-          ) : null,
-          borderRadius: BorderRadius.circular(widget.borderRadius),
-        );
-        return GestureDetector(
-          //Stack best describes the relation between the animation and the container box
-          child: AnimatedBuilder(
-            animation: _shrinkController,
-            builder: (context, child) {
-              var _shrinkAnimation = Tween<double>(begin: 1, end: 0.9).animate(CurvedAnimation(
-                parent: _shrinkController,
-                curve: Curves.elasticOut,
-                reverseCurve: Curves.elasticIn,
-              ));
-              return Transform.scale(
-                scale: _shrinkAnimation.value,
-                alignment: Alignment.center,
-                child: child,
-              );
-            },
-            child: Stack(
-              alignment: Alignment.center,
-              children: [
-                //Container
-                Positioned.fill(
-                  child: AnimatedContainer(
-                    duration: const Duration(milliseconds: 750),
-                    decoration: decoration,
-                    alignment: Alignment.center,
-                  ),
-                ),
-                //Animation ( similar to ink well effect )
-                Positioned.fill(
-                  child: AnimatedBuilder(
-                    animation: _scaleController,
-                    child: AnimatedBuilder(
-                      animation: _opacController,
-                      builder: (context, child) {
-                        num topLeftSquredDis = pow(pressedX, 2) + pow(pressedY, 2);
-                        num topRightSquredDis = pow(constraints.maxWidth - pressedX, 2) + pow(pressedY, 2);
-                        num botLeftSquredDis = pow(pressedX, 2) + pow(constraints.maxHeight - pressedY, 2);
-                        num botRightSquredDis = pow(constraints.maxWidth - pressedX, 2) + pow(constraints.maxHeight - pressedY, 2);
-                        num radius = sqrt(max(max(topLeftSquredDis, topRightSquredDis), max(botLeftSquredDis, botRightSquredDis)));
-                        var _opacityAnimation = Tween<double>(begin: 1, end: 0).animate(_opacController);
-                        return OverflowBox(
-                              maxHeight: radius * 2,
-                              maxWidth: radius * 2,
-                              child: SizedBox(
-                                child: Opacity(
-                                  opacity: _opacityAnimation.value,
-                                  child: Container(
-                                    decoration: BoxDecoration(
-                                      shape: BoxShape.circle,
-                                      color: widget.selectedBackgroundColor,
-                                    ),
-                                  ),
-                                ),
-                                width: radius * 2,
-                                height: radius * 2,
-                              ),
-                        );
-                      },
-                    ),
-                    builder: (context, child) {
-                      var _scaleAnimation = Tween<double>(begin: 0, end: 1).animate(CurvedAnimation(
-                        parent: _scaleController,
-                        curve: Curves.easeInOutCubic,
-                      ));
-                      return ClipRRect(
-                        child: Transform.translate(
-                          child: Transform.scale(
-                            scale: _scaleAnimation.value,
-                            child: child,
-                          ),
-                          offset: Offset(pressedX - constraints.maxWidth / 2, pressedY - constraints.maxHeight / 2),
-                        ),
-                        clipBehavior: Clip.antiAlias,
-                        borderRadius: BorderRadius.circular(widget.borderRadius),
-                      );
-                    },
-                  )
-                ),
-                Positioned(
-                  child: AnimatedSwitcher(
-                    duration: const Duration(milliseconds: 750),
-                    child: data.selected == widget.id ? widget.selectedChild : widget.unselectedChild,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          onTapUp: (detail) {
-            setState(() {
-              selected = true;
-              //Notify the group manager
-              SelectNotification(selection: widget.id).dispatch(context);
-              //Record the coordinate and start the animation
-              pressedX = detail.localPosition.dx;
-              pressedY = detail.localPosition.dy;
-              _opacController.reset();
-              _scaleController.reset();
-              _scaleController.forward();
-              _shrinkController.reverse();
-            });
-          },
-          onTapDown: ((detail) {
-            _shrinkController.value = _shrinkController.upperBound;
-          }),
-          onTapCancel: (() {
+          )
+        ],
+        color: widget.unselectedBackgroundColor,
+        border: widget.borderWidth > 0
+            ? Border.all(color: data.selected == widget.id ? widget.selectedBackgroundColor : widget.borderColor, width: widget.borderWidth)
+            : null,
+        borderRadius: BorderRadius.circular(widget.borderRadius),
+      );
+      return GestureDetector(
+        //Stack best describes the relation between the animation and the container box
+        onTapUp: (detail) {
+          setState(() {
+            selected = true;
+            //Notify the group manager
+            SelectNotification(selection: widget.id).dispatch(context);
+            //Record the coordinate and start the animation
+            pressedX = detail.localPosition.dx;
+            pressedY = detail.localPosition.dy;
+            _opacController.reset();
+            _scaleController.reset();
+            _scaleController.forward();
             _shrinkController.reverse();
-          }),
-        );
-      }
-    );
+          });
+        },
+        onTapDown: ((detail) {
+          _shrinkController.value = _shrinkController.upperBound;
+        }),
+        onTapCancel: (() {
+          _shrinkController.reverse();
+        }),
+        //Stack best describes the relation between the animation and the container box
+        child: AnimatedBuilder(
+          animation: _shrinkController,
+          builder: (context, child) {
+            var shrinkAnimation = Tween<double>(begin: 1, end: 0.9).animate(CurvedAnimation(
+              parent: _shrinkController,
+              curve: Curves.elasticOut,
+              reverseCurve: Curves.elasticIn,
+            ));
+            return Transform.scale(
+              scale: shrinkAnimation.value,
+              alignment: Alignment.center,
+              child: child,
+            );
+          },
+          child: Stack(
+            alignment: Alignment.center,
+            children: [
+              //Container
+              Positioned.fill(
+                child: AnimatedContainer(
+                  duration: const Duration(milliseconds: 300),
+                  decoration: decoration,
+                  alignment: Alignment.center,
+                ),
+              ),
+              //Animation ( similar to ink well effect )
+              Positioned.fill(
+                  child: AnimatedBuilder(
+                animation: _scaleController,
+                child: AnimatedBuilder(
+                  animation: _opacController,
+                  builder: (context, child) {
+                    num topLeftSquredDis = pow(pressedX, 2) + pow(pressedY, 2);
+                    num topRightSquredDis = pow(constraints.maxWidth - pressedX, 2) + pow(pressedY, 2);
+                    num botLeftSquredDis = pow(pressedX, 2) + pow(constraints.maxHeight - pressedY, 2);
+                    num botRightSquredDis = pow(constraints.maxWidth - pressedX, 2) + pow(constraints.maxHeight - pressedY, 2);
+                    num radius = sqrt(max(max(topLeftSquredDis, topRightSquredDis), max(botLeftSquredDis, botRightSquredDis)));
+                    var opacityAnimation = Tween<double>(begin: 1, end: 0).animate(_opacController);
+                    return OverflowBox(
+                      maxHeight: radius * 2,
+                      maxWidth: radius * 2,
+                      child: SizedBox(
+                        width: radius * 2,
+                        height: radius * 2,
+                        child: Opacity(
+                          opacity: opacityAnimation.value,
+                          child: Container(
+                            decoration: BoxDecoration(
+                              shape: BoxShape.circle,
+                              color: widget.selectedBackgroundColor,
+                            ),
+                          ),
+                        ),
+                      ),
+                    );
+                  },
+                ),
+                builder: (context, child) {
+                  var scaleAnimation = Tween<double>(begin: 0, end: 1).animate(CurvedAnimation(
+                    parent: _scaleController,
+                    curve: Curves.easeInOutCubic,
+                  ));
+                  return ClipRRect(
+                    clipBehavior: Clip.antiAlias,
+                    borderRadius: BorderRadius.circular(widget.borderRadius),
+                    child: Transform.translate(
+                      offset: Offset(pressedX - constraints.maxWidth / 2, pressedY - constraints.maxHeight / 2),
+                      child: Transform.scale(
+                        scale: scaleAnimation.value,
+                        child: child,
+                      ),
+                    ),
+                  );
+                },
+              )),
+              Positioned(
+                child: AnimatedSwitcher(
+                  duration: const Duration(milliseconds: 300),
+                  child: data.selected == widget.id ? widget.selectedChild : widget.unselectedChild,
+                ),
+              ),
+            ],
+          ),
+        ),
+      );
+    });
   }
 }
 
@@ -260,32 +258,39 @@ Widget generateSelectionBox({
                   children: [
                     Padding(
                       padding: const EdgeInsets.symmetric(vertical: 10.0),
-                      child: Text(primaryText, style: TextStyle(
-                        color: Colors.grey[850],
-                        fontWeight: FontWeight.bold,
-                        fontSize: primaryTextSize,
-                      ),),
+                      child: Text(
+                        primaryText,
+                        style: TextStyle(
+                          color: Colors.grey[850],
+                          fontWeight: FontWeight.bold,
+                          fontSize: primaryTextSize,
+                        ),
+                      ),
                     ),
-                    if(secondaryText != null)
+                    if (secondaryText != null)
                       Padding(
                         padding: const EdgeInsets.symmetric(vertical: 10.0),
-                        child: Text(secondaryText, style: TextStyle(
-                          color: Colors.grey[850]!.withOpacity(0.8),
-                          fontSize: secondaryTextSize,
-                        ),),
+                        child: Text(
+                          secondaryText,
+                          style: TextStyle(
+                            color: Colors.grey[850]!.withOpacity(0.8),
+                            fontSize: secondaryTextSize,
+                          ),
+                        ),
                       ),
                   ],
                 ),
               ),
             ),
           ),
-          if(decorationText != null)
+          if (decorationText != null)
             Positioned.fill(
-              child: ClipRRect(
-                child: Align(
-                  alignment: Alignment.bottomRight,
-                  child: Transform.translate(
-                    offset: const Offset(15, 35), 
+                child: ClipRRect(
+              borderRadius: BorderRadius.circular(10.0),
+              child: Align(
+                alignment: Alignment.bottomRight,
+                child: Transform.translate(
+                    offset: const Offset(15, 35),
                     child: Text(
                       decorationText,
                       style: TextStyle(
@@ -293,12 +298,9 @@ Widget generateSelectionBox({
                         fontSize: 120.0,
                         fontWeight: FontWeight.bold,
                       ),
-                    )
-                  ),
-                ),
-                borderRadius: BorderRadius.circular(10.0),
-              )
-            )
+                    )),
+              ),
+            ))
         ],
       ),
       selectedChild: Stack(
@@ -316,32 +318,39 @@ Widget generateSelectionBox({
                   children: [
                     Padding(
                       padding: const EdgeInsets.symmetric(vertical: 10.0),
-                      child: Text(primaryText, style: TextStyle(
-                        color: Colors.white,
-                        fontWeight: FontWeight.bold,
-                        fontSize: primaryTextSize,
-                      ),),
+                      child: Text(
+                        primaryText,
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontWeight: FontWeight.bold,
+                          fontSize: primaryTextSize,
+                        ),
+                      ),
                     ),
-                    if(secondaryText != null)
+                    if (secondaryText != null)
                       Padding(
                         padding: const EdgeInsets.symmetric(vertical: 10.0),
-                        child: Text(secondaryText, style: TextStyle(
-                          color: Colors.white.withOpacity(0.8),
-                          fontSize: secondaryTextSize,
-                        ),),
+                        child: Text(
+                          secondaryText,
+                          style: TextStyle(
+                            color: Colors.white.withOpacity(0.8),
+                            fontSize: secondaryTextSize,
+                          ),
+                        ),
                       ),
                   ],
                 ),
               ),
             ),
           ),
-          if(decorationText != null)
+          if (decorationText != null)
             Positioned.fill(
-              child: ClipRRect(
-                child: Align(
-                  alignment: Alignment.bottomRight,
-                  child: Transform.translate(
-                    offset: const Offset(15, 35), 
+                child: ClipRRect(
+              borderRadius: BorderRadius.circular(10.0),
+              child: Align(
+                alignment: Alignment.bottomRight,
+                child: Transform.translate(
+                    offset: const Offset(15, 35),
                     child: Text(
                       decorationText,
                       style: TextStyle(
@@ -349,12 +358,9 @@ Widget generateSelectionBox({
                         fontSize: 120.0,
                         fontWeight: FontWeight.bold,
                       ),
-                    )
-                  ),
-                ),
-                borderRadius: BorderRadius.circular(10.0),
-              )
-            )
+                    )),
+              ),
+            ))
         ],
       ),
     ),
